Affichage sorties 4

+++ main.py
@@ -29,7 +29,7 @@ if __name__ == '__main__':
     cloudy_image = splicing_functions.convert_float32_to_uint8(cloudy_image)
     background = splicing_functions.convert_float32_to_uint8(background)
     source = splicing_functions.convert_float32_to_uint8(source)
-    mask = int(mask * 255)
+    mask = mask * 255
 
     tifffile.imwrite('output/cloudy.png', cloudy_image)
     tifffile.imwrite('output/background.png', background)
